+++ src/table.cpp
@@ -1,0 +1,6 @@
+
+#include "table_impl.h"
+
+namespace calicodb {
+
+} // namespace calicodb
